+++ apps/website/CHANGELOG.md
@@ -1,5 +1,12 @@
 # website
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - negative-space@2.4.3
+
 ## 0.0.3
 
 ### Patch Changes